refactor: move standard package terms out of the solve loop
The terms of the standard packages were formed inline in two files that had
grown past fifteen hundred lines each, alongside the solve they feed. They are
now in a packages directory beside the advanced ones, which already had theirs:

  - storage: the terms coupling one time step to the next
  - npf: the conductivity derivatives and the saturation they follow
  - head_dependent: the boundary derivatives, and the corner a drain has,
    which boundary.py already held on its own
  - recharge: the flow a recharge value produces

Nothing changed but where the code sits. Each moved function reached only for
the API handle and the model name, so those are arguments now rather than
attributes, which is the shape the advanced packages already use. Every name
moved was private, so the package interface is unchanged.
