+++ mf6adj/packages/__init__.py
@@ -1,0 +1,1 @@
+"""Standard MODFLOW 6 packages whose terms the adjoint solution forms."""
